echoctl: Add "poll" command to query I/O status
This command uses the poll(2) system call to determine the current I/O
status of the echo device.  The request can be limited to checking
either read or write status but defaults to checking both.  If the
device is readable, the FIONREAD ioctl is used to determine how many
bytes can be read.  If the device is writable, the FIONWRITE ioctl is
used to determine how much room is available for writing.