echoctl: Add "events" command to report kernel events
This command uses the kqueue(2) mechanism to register EVFILT_READ and
EVFILT_WRITE events for the echo device.  As with the "poll" command,
this command can be limited to only checking for read or write, and to
either do a single poll of device state or wait in a loop reporting
events.  By default it does a single poll for both events.

Since these events are level-triggered by default, this command
registers the events with EV_CLEAR so that new events are only
reported when the driver reports a state change.