mappage: Use a physical memory VM object with the mmap_single method
The single page of wired memory is allocated on the first fault by the
physical pager.  Since the physical memory pager frees its pages on
destruction, the driver can be safely unloaded.  The wired memory will
be freed once the last mapping is deleted after the driver has been
unloaded.