mappage: Use the extended device pager
Revert back to exposing a wired page allocated via malloc(9), but use
the extended device pager.  The driver can be unloaded if there are no
active mappings.  If there are active mappings of the wired page,
module unloading fails with EBUSY.  The device pager operation
constructor and destructor are used to track when there are active
mappings.